* commit (not all change of develop but compiller)

+++ MediaShop.Common/Interfaces/Repositories/IRepository.cs
@@ -21,7 +21,7 @@ namespace MediaShop.Common.Interfaces.Repositories
         /// </summary>
         /// <param name="id">Id of searching element</param>
         /// <returns>TModel</returns>
-        TModel Get(int id);
+        TModel Get(long id);
 
         /// <summary>
         /// interface method add for type TModel
@@ -49,7 +49,7 @@ namespace MediaShop.Common.Interfaces.Repositories
         /// </summary>
         /// <param name="id">Id of deleting model</param>
         /// <returns>Deleted model</returns>
-        TModel Delete(int id);
+        TModel Delete(long id);
 
         /// <summary>
         /// Find by id
